pyproject.toml
[tool.poetry]
name = "auto_web"
version = "0.1.0"
description = "Automation tool for Rexpeita operations"
authors = ["Lucas Picinini <137008535+lucaspicinini@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
attrs = "25.3.0"
certifi = "2025.1.31"
h11 = "0.14.0"
idna = "3.10"
outcome = "1.3.0.post0"
PySocks = "1.7.1"
python-dotenv = "1.1.0"
selenium = "4.31.0"
sniffio = "1.3.1"
sortedcontainers = "2.4.0"
trio = "0.29.0"
trio-websocket = "0.12.0"
typing-extensions = "4.13.1"
urllib3 = "2.3.0"
websocket-client = "1.8.0"
wsproto = "1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
